deps: update V8 to 9.3.345.16

+++ deps/v8/src/wasm/wasm-external-refs.cc
@@ -238,7 +238,7 @@ void float32_to_uint64_sat_wrapper(Address data) {
     WriteUnalignedValue<uint64_t>(data, static_cast<uint64_t>(input));
     return;
   }
-  if (input >= std::numeric_limits<uint64_t>::max()) {
+  if (input >= static_cast<float>(std::numeric_limits<uint64_t>::max())) {
     WriteUnalignedValue<uint64_t>(data, std::numeric_limits<uint64_t>::max());
     return;
   }
@@ -268,7 +268,7 @@ void float64_to_uint64_sat_wrapper(Address data) {
     WriteUnalignedValue<uint64_t>(data, static_cast<uint64_t>(input));
     return;
   }
-  if (input >= std::numeric_limits<uint64_t>::max()) {
+  if (input >= static_cast<double>(std::numeric_limits<uint64_t>::max())) {
     WriteUnalignedValue<uint64_t>(data, std::numeric_limits<uint64_t>::max());
     return;
   }
